Add httpoison to applications in mix.exs.
Otherwise exometer_datadog might fail on startup, because httpoison (and
therefore hackney) have not yet been started.

Had to add HTTPoison as a dependency as well - this wasn't previously
the case, as I wanted people to be able to swap in http clients at will.
I've marked it as optional, so should be easy enough to remove?