Resolve env vars before deciding to start reporter.
ExometerDatadog checks whether it has an API key.  If the key is missing
it'll avoid registering the reporter, and creating any metrics.

We also allow users to tell ExometerDatadog to get the api key from
environment variables.  However, due to the way the API key check was
made, it assumed that these env variables would always be present.

This means that any application that uses ExometerDatadog and is
configured to pull in env vars would crash if the env vars weren't
actually defined.

This updates the check to resolve env vars first.