HTTPoison is no longer an optional dependency.
Ideally I wanted httpoison to be an optional dependency so people could
swap in their own http clients as they wanted.

However, I misunderstood the purpose of optional dependencies a little -
they need to be explicitly included in client applications mix.exs in
order to be installed.

This on it's own isn't too bad, but it means we can't include httpoison
in our `applications` list in mix.exs either.  We could easily just start
this application ourselves if the code is present, but then clients need
to remember to add httpoison to their applications list in order for OTP
releases to work properly.

All in all, I'm not desperate enough for httpoison to be an optional
dependency to go though all the hassle of documenting this.  If anyone
asks for httpoison to be optional I can easily revert.